backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
google-generativeai==0.8.4
requests==2.32.3
openpyxl==3.1.5
gunicorn==23.0.0
